move sts package to its own repo
