--- tests/system/sts/README.md
@@ -1,5 +1,0 @@
-# System Test Service
-Simple web service for systems testing. 
-
-## Update Image
-Run `make release`
