--- tests/system/sts/api/api/Dockerrun.aws.json
@@ -1,17 +1,0 @@
-{
-    "AWSEBDockerrunVersion": 2,
-    "containerDefinitions": [
-        {
-            "name": "sts",
-            "image": "quintilesims/sts",
-            "essential": true,
-            "memory": 128,
-            "portMappings": [
-                {
-                    "hostPort": 80,
-                    "containerPort": 80
-                }
-            ]
-        }
-    ]
-}
