--- tests/system/sts/api/api/Makefile
@@ -1,7 +1,0 @@
-DOCKER_IMAGE = "quintilesims/sts"
-
-release:
-	docker build -t $(DOCKER_IMAGE) .
-	docker push $(DOCKER_IMAGE)
-
-.PHONY: release
